requirements.txt
absl-py==1.4.0
altgraph==0.17.3
astunparse==1.6.3
attrs==22.2.0
auto-py-to-exe==2.36.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.27.0
contourpy==1.0.7
cycler==0.11.0
deptry==0.12.0
Eel==0.16.0
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lit==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mediapipe==0.10.2
ml-dtypes==0.1.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.2
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
pyqt5-tools==5.15.9.3.3
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.5
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
sympy==1.12
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
whichcraft==0.6.1
wrapt==1.14.1
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
